:recycle: [client] runTwoPartyTest: Move from happy_test to client_role_test
As it is used not just in happy_test.
