+++ client/client_role_test.go
@@ -15,20 +15,29 @@
 package client_test
 
 import (
+	"context"
+	"math/big"
 	"math/rand"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 
+	"perun.network/go-perun/apps/payment"
+	chtest "perun.network/go-perun/channel/test"
 	"perun.network/go-perun/client"
 	ctest "perun.network/go-perun/client/test"
+	"perun.network/go-perun/pkg/test"
 	wtest "perun.network/go-perun/wallet/test"
 	"perun.network/go-perun/watcher/local"
+	"perun.network/go-perun/wire"
 	wiretest "perun.network/go-perun/wire/test"
 )
 
-const roleOperationTimeout = 1 * time.Second
+const (
+	roleOperationTimeout = 1 * time.Second
+	twoPartyTestTimeout  = 10 * time.Second
+)
 
 func NewSetups(rng *rand.Rand, names []string) []ctest.RoleSetup {
 	var (
@@ -78,4 +87,31 @@ func NewClients(rng *rand.Rand, names []string, t *testing.T) []*Client {
 		}
 	}
 	return clients
+}
+
+func runTwoPartyTest(ctx context.Context, t *testing.T, setup func(*rand.Rand) ([]ctest.RoleSetup, [2]ctest.Executer)) {
+	rng := test.Prng(t)
+	for i := 0; i < 2; i++ {
+		setups, roles := setup(rng)
+		app := client.WithoutApp()
+		if i == 1 {
+			app = client.WithApp(
+				chtest.NewRandomAppAndData(rng, chtest.WithAppRandomizer(new(payment.Randomizer))),
+			)
+		}
+
+		cfg := &ctest.AliceBobExecConfig{
+			BaseExecConfig: ctest.MakeBaseExecConfig(
+				[2]wire.Address{setups[0].Identity.Address(), setups[1].Identity.Address()},
+				chtest.NewRandomAsset(rng),
+				[2]*big.Int{big.NewInt(100), big.NewInt(100)},
+				app,
+			),
+			NumPayments: [2]int{2, 2},
+			TxAmounts:   [2]*big.Int{big.NewInt(5), big.NewInt(3)},
+		}
+
+		err := ctest.ExecuteTwoPartyTest(ctx, roles, cfg)
+		assert.NoError(t, err)
+	}
 }
